coperniq-mep-templates/sandbox/requirements.txt
# Coperniq MEP Templates - Python Dependencies
# Instance 388 (Kipper Energy Solutions)

# =============================================================================
# Core LangChain/LangGraph (NO OPENAI)
# =============================================================================
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0  # Claude models
langserve>=0.3.0            # Chat UI

# =============================================================================
# HTTP & API
# =============================================================================
httpx>=0.27.0               # Async HTTP client
fastapi>=0.115.0            # API framework
uvicorn>=0.32.0             # ASGI server
pydantic>=2.0.0             # Data validation

# =============================================================================
# Voice AI (Optional - for voice agents)
# =============================================================================
# deepgram-sdk>=3.0.0       # STT - Deepgram Nova-3
# cartesia>=0.4.0           # TTS - Cartesia Sonic-3
# twilio>=9.0.0             # Phone integration
# websockets>=12.0          # Audio streaming

# =============================================================================
# Observability
# =============================================================================
langsmith>=0.1.0            # LangSmith tracing

# =============================================================================
# Development
# =============================================================================
python-dotenv>=1.0.0        # Environment variables
rich>=13.0.0                # CLI formatting (optional)
typer>=0.12.0               # CLI framework (optional)
